fix: change integer to double for product price + value type for enum columns

+++ inventory-service/src/main/java/camellia/ecommerce/inventory_service/kafka/events/ProductEvent.java
@@ -4,6 +4,6 @@ import java.util.UUID;
 
 import camellia.ecommerce.inventory_service.enums.Category;
 
-public record ProductEvent(UUID publicId, String name, Integer numberInStock, Integer price, Category category) {
+public record ProductEvent(UUID publicId, String name, Integer numberInStock, Double price, Category category) {
 
 }
